updated content collections to new api style
