+++ src/components/NotePreviewCard.astro
@@ -13,7 +13,7 @@ const dateString = new Date(dateUpdated).toLocaleDateString("en-us", {
   dateStyle: "medium",
 });
 const addFontClamp = title.length > 45;
-const postUrl = `/notes/${noteInfo.slug}`;
+const postUrl = `/notes/${noteInfo.id}`;
 ---
 
 <CardWrapper>
