+++ src/content/blog/2021-01-08-building-snake.mdx
@@ -3,7 +3,6 @@ title: Building Snake with HTML Canvas and JS
 description: I made my own version of Snake! Here's what I learned in the process.
 publishedDate: 2021-01-08
 featured: false
-layout: ../../layouts/BlogPostLayout.astro
 slug: building-snake-html-canvas-js
 tags:
   - JavaScript
