+++ src/content/blog/2021-10-19-how-to-make-your-first-pull-request-for-hacktoberfest.mdx
@@ -4,7 +4,6 @@ featured: false
 publishedDate: 2021-10-19
 status: Draft
 slug: pull-request-hacktoberfest
-layout: ../../layouts/BlogPostLayout.astro
 type: blog
 description: Leaves are changing color. Pumpkin spice flavors are everywhere. It must be October in the US! This means it's also time for Hacktoberfest!
 tags:
