+++ src/content/blog/2022-01-04-content-projection-in-angular.mdx
@@ -4,7 +4,6 @@ featured: false
 publishedDate: 2022-01-04
 status: Draft
 slug: content-projection-angular
-layout: ../../layouts/BlogPostLayout.astro
 type: blog
 description: We'll go over what content projection looks like and how you can use it to make components that are reusable and less tied to the data they display.
 tags:
