+++ src/content/blog/2022-03-10-international-women's-day-recap.mdx
@@ -4,7 +4,6 @@ featured: false
 publishedDate: 2022-03-10
 status: Draft
 slug: international-women-day-recap
-layout: ../../layouts/BlogPostLayout.astro
 type: blog
 description: "We hosted a live event with Women Techmakers, featuring talks and a panel discussion on this year's topic: progress over perfection."
 tags:
