+++ src/content/blog/2022-11-18-how-to-create-a-custom-login-command-with-cypress,-react,-&-auth0.mdx
@@ -4,7 +4,6 @@ featured: false
 publishedDate: 2022-11-18
 status: Draft
 slug: create-custom-login-command-cypress-react-auth0
-layout: ../../layouts/BlogPostLayout.astro
 type: blog
 description: We’ll go over the steps you’ll need to implement so you can have a custom Cypress command to log in a user and let your front end code continue to work safely!
 tags:
