+++ src/content/blog/2023-01-06-leveling-up-your-work-through-architecture-design-and-time-estimation.mdx
@@ -4,7 +4,6 @@ featured: false
 publishedDate: 2023-01-06
 status: Draft
 slug: leveling-work-architecture-design-time-estimation
-layout: ../../layouts/BlogPostLayout.astro
 type: blog
 description: Help deepen your understanding of the software you’re writing by improving your ability to estimate your time and discuss your technical decisions.
 tags:
