+++ src/content/blog/2023-07-17-introducing-the-vue-3-and-xstate-kit-for-starter.dev.mdx
@@ -3,7 +3,6 @@ title: Introducing the Vue 3 and XState kit for starter.dev
 featured: false
 publishedDate: 2023-07-17
 slug: introducing-vue-3-xstate-kit-starter-dev
-layout: ../../layouts/BlogPostLayout.astro
 type: blog
 remoteLink: https://www.thisdot.co/blog/introducing-the-vue-3-and-xstate-kit-for-starter-dev
 remoteSrc: This Dot
