+++ src/content/blog/2023-11-07-an-origin-story.mdx
@@ -2,7 +2,6 @@
 title: An Origin Story
 description: A quick walkthrough of my history, then thoughts on where I'm at today and where I want to go.
 featured: true
-layout: ../../layouts/BlogPostLayout.astro
 slug: origin-story
 publishedDate: 2023-11-08
 tags:
