+++ src/content/blog/2023-12-02-code-breaker-now-live-in-the-arcade.mdx
@@ -1,7 +1,6 @@
 ---
 title: Code Breaker now live in the Arcade!
 featured: false
-layout: ../../layouts/BlogPostLayout.astro
 slug: code-breaker-live-arcade
 publishedDate: 2023-12-02
 description: Play the new Code Breaker game on my arcade site now!
