+++ src/content/blog/2024-01-01-2023-year-in-review.mdx
@@ -1,7 +1,6 @@
 ---
 title: 2023 Year in Review
 featured: false
-layout: ../../layouts/BlogPostLayout.astro
 slug: 2023-year-review
 publishedDate: 2024-01-01
 tags:
