+++ src/content/blog/2024-02-07-my-that-experience.mdx
@@ -1,7 +1,6 @@
 ---
 title: My THAT Conference Experience
 featured: true
-layout: ../../layouts/BlogPostLayout.astro
 slug: my-that-experience
 publishedDate: 2024-02-07
 description: My favorite moments and takeaways from THAT Conference!
