+++ src/content/blog/2024-12-31-2024-year-in-review.mdx
@@ -1,7 +1,6 @@
 ---
 title: 2024 Year in Review
 featured: true
-layout: ../../layouts/BlogPostLayout.astro
 slug: 2024-year-in-review
 publishedDate: 2024-12-31
 tags:
