+++ src/content/notes/2023-12-01-beginner-javascript.mdx
@@ -1,7 +1,6 @@
 ---
 title: Beginner JavaScript
 description: Notes on Wes Bos' Beginner JS course
-layout: ../../layouts/BlogPostLayout.astro
 slug: beginner-javascript
 status: Garden
 dateUpdated: 2020-08-20
