+++ src/content/notes/2023-12-01-css-snippets.mdx
@@ -1,7 +1,6 @@
 ---
 title: CSS Snippets
 description: My own CSS reference cheat sheet, when I can't remember how to do things
-layout: ../../layouts/BlogPostLayout.astro
 slug: css-snippets
 dateUpdated: 2020-08-19
 tags:
