+++ src/content/notes/2023-12-01-effective-typescript.mdx
@@ -1,7 +1,6 @@
 ---
 title: Effective Typescript
 description: Notes and impressions from reading through Effective TypeScript by Dan Vanderkam
-layout: ../../layouts/BlogPostLayout.astro
 slug: effective-typescript
 dateUpdated: 2022-04-26
 source: Dan Vanderkam
