+++ src/content/notes/2023-12-01-git-snippets.mdx
@@ -1,7 +1,6 @@
 ---
 title: Git Snippets
 description: A reference for the less used Git commands, so I can find them easily
-layout: ../../layouts/BlogPostLayout.astro
 slug: git-snippets
 dateUpdated: 2022-01-13
 tags:
