+++ src/content/notes/2023-12-01-javascript-snippets.mdx
@@ -1,7 +1,6 @@
 ---
 title: JavaScript Snippets
 description: My own reference sheet for common, basic JS recipes that I use and/or need to look up often
-layout: ../../layouts/BlogPostLayout.astro
 slug: javascript-snippets
 dateUpdated: 2020-08-11
 tags:
