+++ src/content/notes/2023-12-01-project-snippets.mdx
@@ -1,7 +1,6 @@
 ---
 title: Project Snippets
 description: Helpful references for getting started on projects
-layout: ../../layouts/BlogPostLayout.astro
 slug: project-snippets
 dateUpdated: 2021-05-07
 tags:
