+++ src/content/notes/2023-12-01-technical-interviews.mdx
@@ -1,7 +1,6 @@
 ---
 title: Technical Interviews
 description: Notes while reading through "De-coding the Technical Interview" by Emma Bostian
-layout: ../../layouts/BlogPostLayout.astro
 slug: technical-interviews
 dateUpdated: 2021-04-12
 source: Emma Bostian
