+++ src/content/notes/2023-12-01-typescript-basics.mdx
@@ -1,7 +1,6 @@
 ---
 title: TypeScript Basics
 description: Notes on the basics of the TypeScript language
-layout: ../../layouts/BlogPostLayout.astro
 slug: typescript-basics
 dateUpdated: 2021-10-26
 source: TypeScript Docs
