+++ src/content/projects/_project-template.mdx
@@ -12,7 +12,6 @@
   codeLink: null
   liveLink: null
   hasCaseStudy: true
-  layout: ../../layouts/CaseStudyLayout.astro
 ---
 
 import CaseStudyStats from "@components/CaseStudyStats.astro";
