+++ src/content/projects/arcade.mdx
@@ -12,7 +12,6 @@
   codeLink: "https://github.com/lindakatcodes/arcade"
   liveLink: "https://lindakat-arcade.netlify.app/"
   hasCaseStudy: true
-  layout: ../../layouts/CaseStudyLayout.astro
 ---
 import CaseStudyStats from "@components/CaseStudyStats.astro";
 import { Image } from "astro:assets";
