+++ src/content/projects/do-i-own-this.mdx
@@ -12,7 +12,6 @@
   codeLink: "https://github.com/lindakatcodes/doIOwnThis"
   liveLink: "https://doiownthis.netlify.app/"
   hasCaseStudy: true
-  layout: ../../layouts/CaseStudyLayout.astro
 ---
 
 import CaseStudyStats from "@components/CaseStudyStats.astro";
