+++ src/pages/blog/[...slug].astro
@@ -3,20 +3,21 @@ import PostDetails from "@components/PostDetails.astro";
 import PaginationLinks from "@components/PaginationLinks.astro";
 import TableOfContents from "@components/TableOfContents.astro";
 import MainLayout from "@layouts/MainLayout.astro";
-import { getCollection } from "astro:content";
+import BlogPostLayout from "@layouts/BlogPostLayout.astro";
+import { getCollection, render } from "astro:content";
 
 export async function getStaticPaths() {
   const blogPosts = await getCollection("blog", ({ data }) => {
     return data.remoteLink === undefined;
   });
   return blogPosts.map((entry) => ({
-    params: { slug: entry.slug },
+    params: { slug: entry.id },
     props: { entry },
   }));
 }
 
 const { entry } = Astro.props;
-const { Content, headings, remarkPluginFrontmatter } = await entry.render();
+const { Content, headings, remarkPluginFrontmatter } = await render(entry);
 const canonicalUrl = new URL(Astro.url.pathname, Astro.site);
 
 const metadata = {
@@ -50,7 +51,9 @@ const postDetails = {
       <TableOfContents headings={headings} />
     </div>
     <div class="body">
-      <Content />
+      <BlogPostLayout>
+        <Content />
+      </BlogPostLayout>
     </div>
     <div class="links">
       <PaginationLinks main="#top" mainText="Back to top" />
